% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MCpower}
\alias{MCpower}
\title{MC power for selecting the best regime in SMART}
\usage{
MCpower(
  Q = c(0.5, 0.5, 0.5),
  pi1 = c(0.4, 0.35, 0.25),
  pi2 = c(0.35, 0.32, 0.36, 0.56, 0.37, 0.4),
  P = c(1/3, 1/3, 1/3),
  n = 365,
  m = 10000,
  seed = 2021
)
}
\arguments{
\item{Q}{a vector of (Q12, Q21, Q31) for second stage randomization probability in SMART}

\item{pi1}{the respond rate of treatment 1, 2, 3 in stage I in SMART}

\item{pi2}{the respond rate for patients who do not respond to a treatment in stage I, but
respond to another treatment in stage II}

\item{P}{Randomization probability in stage I}

\item{n}{Sample size in SMART}

\item{m}{number of Monte Carlo repetitions}

\item{seed}{seed for generate MC repetitions}
}
\description{
MC power for selecting the best regime in SMART
}
\examples{

seed = 2021
Q = c(0.5, 0.5, 0.5)
pi1 = c(0.4, 0.35, 0.25)
pi2 = c(0.35, 0.32, 0.36, 0.56, 0.37, 0.4)
P = c(1/3, 1/3, 1/3)
n = 365
m = 10000

MCpower(Q, pi1, pi2, P, n, m, seed)

}
