% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mfc.default}
\alias{mfc.default}
\title{minimum failure count in RASMART subject to a lower bound of the power}
\usage{
\method{mfc}{default}(x = c(0.5, 0.5, 0.5), pi1 = c(0.4, 0.35, 0.25),
  pi2 = c(0.35, 0.32, 0.36, 0.56, 0.37, 0.4), P = c(1/3, 1/3, 1/3),
  n = 365, power = 0.8, alpha.step = 0.8, lambda_seq = seq(1, 100,
  by = 0.5), ...)
}
\arguments{
\item{x}{initial second stage randomization probability in SMART}

\item{pi1}{the respond rate of treatment 1, 2, 3 in stage I in SMART}

\item{pi2}{the respond rate for patients who do not respond to a treatment in stage I, but
respond to another treatment in stage II}

\item{P}{Randomization probability in stage I}

\item{n}{Sample size in SMART}

\item{power}{the minimum controlled power for RASMART}

\item{alpha.step}{The step size to search a small alpha for satisfying the Lipschitz constant
such that f(x) <= f(x_0) + df(x_0)(x-x_0) + 1/(2alpha)||x-x_0||^2. Default is 0.8.}

\item{lambda_seq}{The penalty parameter to minimize the lagrangian function.}

\item{...}{other arguments related to mfc}
}
\description{
minimum failure count in RASMART subject to a lower bound of the power
}
\examples{

x = c(0.3, 0.3, 0.3)
pi1 = c(0.4, 0.35, 0.25)
pi2 = c(0.35, 0.32, 0.36, 0.56, 0.37, 0.4)
P = c(1/3, 1/3, 1/3)
n = 365
power = 0.8
alpha.step = 0.8
lambda_seq = seq(1,50, by = 1)

mfc(x, pi1, pi2, P, n, power, alpha.step, lambda_seq)

}
