% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mfc}
\alias{mfc}
\alias{print.mfc}
\title{mfc generic}
\usage{
mfc(x, ...)

\method{print}{mfc}(x, ...)
}
\arguments{
\item{x}{an object of class from ''mfc''.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
mfc generic
}
\seealso{
\code{\link{mfc.default}}
}
\keyword{internal}
