% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DSpower.R
\name{plot_DSpower}
\alias{plot_DSpower}
\title{2D or 3D visualization for the Dunn-Sidak power}
\usage{
plot_DSpower(Q1, Q2, Q3, pi1, pi2, P, n, ...)
}
\arguments{
\item{Q1}{a vector of numeric value for the range of Q12 probability}

\item{Q2}{a vector of numeric value for the range of Q21 probability}

\item{Q3}{a vector of numeric value for the range of Q31 probability}

\item{pi1}{the respond rate of treatment 1, 2, 3 in stage I in SMART}

\item{pi2}{the respond rate for patients who do not respond to a treatment in stage I, but
respond to another treatment in stage II}

\item{P}{Randomization probability in stage I}

\item{n}{Sample size in SMART}

\item{...}{other arguments fits to DSpower and plot3D}
}
\value{
a 2D contour plot or 3D plot for the Dunn-Sidak power
}
\description{
2D or 3D visualization for the Dunn-Sidak power
}
\details{
One of Q1, Q2, Q3 must be a single value, and the other two should be a range of
         numerical value between 0 and 1.
}
\examples{

Q1 = 0.3
Q2 = seq(0.1, 0.9, by = 0.005)
Q3 = seq(0.1, 0.9, by = 0.005)
pi1 = c(0.4, 0.35, 0.25)
pi2 = c(0.35, 0.32, 0.36, 0.56, 0.37, 0.4)
P = c(1/3, 1/3, 1/3)
n = 365

plot_DSpower(Q1, Q2, Q3, pi1, pi2, P, n, col = "black", nlevels = 30)

}
